+++ .gitignore
@@ -1,3 +1,4 @@
+emsdk
 vcpkg
 vcpkg_installed
 .pyodide*
