+++ python/core/Cargo.toml
@@ -16,37 +16,53 @@ rust-version = "1.80"
 name = "_rust"
 crate-type = ["cdylib"]
 
+[features]
+default = ["async"]
+async = [
+    "dep:futures",
+    "dep:object_store",
+    "parquet/object_store",
+    "dep:pyo3-asyncio-0-21",
+    "geoarrow/flatgeobuf_async",
+    "geoarrow/parquet_async",
+    "geoarrow/postgis",
+    "dep:tokio",
+]
+
 [dependencies]
 arrow = "52"
 arrow-array = "52"
 arrow-buffer = "52"
 bytes = "1"
 flatgeobuf = { version = "4.2.0", default-features = false }
-futures = "0.3"
-object_store = { version = "0.10", features = ["aws", "azure", "gcp", "http"] }
-parquet = { version = "52", features = ["object_store"] }
+futures = { version = "0.3", optional = true }
+object_store = { version = "0.10", features = [
+    "aws",
+    "azure",
+    "gcp",
+    "http",
+], optional = true }
+parquet = { version = "52" }
 pyo3 = { version = "0.21.0", features = [
     "abi3-py38",
-    "multiple-pymethods",
     "hashbrown",
     "serde",
     "anyhow",
 ] }
-pyo3-arrow = { git = "https://github.com/kylebarron/arro3", rev = "d0d737a03c141ff316e3e354d85828edb42338d4" }
-pyo3-asyncio-0-21 = { version = "0.21", features = ["tokio-runtime"] }
+pyo3-arrow = "0.2"
+pyo3-asyncio-0-21 = { version = "0.21", features = [
+    "tokio-runtime",
+], optional = true }
 pythonize = "0.21"
 geo = "0.28"
 geoarrow = { path = "../../", features = [
     "csv",
-    "flatgeobuf_async",
     "flatgeobuf",
     "geozero",
     "ipc_compression",
-    "parquet_async",
     "parquet_compression",
     "parquet",
     "polylabel",
-    "postgis",
     "rayon",
 ] }
 geozero = { version = "0.13", features = ["with-svg"] }
@@ -54,7 +70,7 @@ numpy = "0.21"
 serde_json = "1"
 sqlx = { version = "0.7", default-features = false, features = ["postgres"] }
 thiserror = "1"
-tokio = { version = "1.9", features = ["rt"] }
+tokio = { version = "1.9", features = ["rt"], optional = true }
 url = "2.5"
 
 # reqwest is pulled in by object store, but not used by python binding itself
