+++ python/core/src/algorithm/geo/area.rs
@@ -1,3 +1,5 @@
+use std::sync::Arc;
+
 use crate::error::PyGeoArrowResult;
 use crate::ffi::from_python::AnyGeometryInput;
 use geoarrow::algorithm::geo::{Area, ChamberlainDuquetteArea, GeodesicArea};
@@ -38,7 +40,7 @@ pub fn area(py: Python, input: AnyGeometryInput, method: AreaMethod) -> PyGeoArr
                 AreaMethod::Euclidean => arr.as_ref().unsigned_area()?,
                 AreaMethod::Geodesic => arr.as_ref().geodesic_area_unsigned()?,
             };
-            Ok(PyArray::from_array(out).to_arro3(py)?)
+            Ok(PyArray::from_array_ref(Arc::new(out)).to_arro3(py)?)
         }
         AnyGeometryInput::Chunked(arr) => {
             let out = match method {
@@ -48,7 +50,7 @@ pub fn area(py: Python, input: AnyGeometryInput, method: AreaMethod) -> PyGeoArr
                 AreaMethod::Euclidean => arr.as_ref().unsigned_area()?,
                 AreaMethod::Geodesic => arr.as_ref().geodesic_area_unsigned()?,
             };
-            Ok(PyChunkedArray::from_arrays(out.chunks())?.to_arro3(py)?)
+            Ok(PyChunkedArray::from_array_refs(out.chunk_refs())?.to_arro3(py)?)
         }
     }
 }
@@ -72,7 +74,7 @@ pub fn signed_area(
                 AreaMethod::Euclidean => arr.as_ref().signed_area()?,
                 AreaMethod::Geodesic => arr.as_ref().geodesic_area_signed()?,
             };
-            Ok(PyArray::from_array(out).to_arro3(py)?)
+            Ok(PyArray::from_array_ref(Arc::new(out)).to_arro3(py)?)
         }
         AnyGeometryInput::Chunked(arr) => {
             let out = match method {
@@ -82,7 +84,7 @@ pub fn signed_area(
                 AreaMethod::Euclidean => arr.as_ref().signed_area()?,
                 AreaMethod::Geodesic => arr.as_ref().geodesic_area_signed()?,
             };
-            Ok(PyChunkedArray::from_arrays(out.chunks())?.to_arro3(py)?)
+            Ok(PyChunkedArray::from_array_refs(out.chunk_refs())?.to_arro3(py)?)
         }
     }
 }
