+++ python/core/src/algorithm/geo/dimensions.rs
@@ -1,3 +1,5 @@
+use std::sync::Arc;
+
 use crate::error::PyGeoArrowResult;
 use crate::ffi::from_python::AnyGeometryInput;
 use geoarrow::algorithm::geo::HasDimensions;
@@ -9,11 +11,11 @@ pub fn is_empty(py: Python, input: AnyGeometryInput) -> PyGeoArrowResult<PyObjec
     match input {
         AnyGeometryInput::Array(arr) => {
             let out = HasDimensions::is_empty(&arr.as_ref())?;
-            Ok(PyArray::from_array(out).to_arro3(py)?)
+            Ok(PyArray::from_array_ref(Arc::new(out)).to_arro3(py)?)
         }
         AnyGeometryInput::Chunked(arr) => {
             let out = HasDimensions::is_empty(&arr.as_ref())?;
-            Ok(PyChunkedArray::from_arrays(out.chunks())?.to_arro3(py)?)
+            Ok(PyChunkedArray::from_array_refs(out.chunk_refs())?.to_arro3(py)?)
         }
     }
 }
