+++ python/core/src/algorithm/geo/frechet_distance.rs
@@ -1,3 +1,5 @@
+use std::sync::Arc;
+
 use crate::error::PyGeoArrowResult;
 use crate::ffi::from_python::input::AnyGeometryBroadcastInput;
 use crate::ffi::from_python::AnyGeometryInput;
@@ -15,21 +17,21 @@ pub fn frechet_distance(
     match (input, other) {
         (AnyGeometryInput::Array(left), AnyGeometryBroadcastInput::Array(right)) => {
             let result = FrechetDistance::frechet_distance(&left.as_ref(), &right.as_ref())?;
-            Ok(PyArray::from_array(result).to_arro3(py)?)
+            Ok(PyArray::from_array_ref(Arc::new(result)).to_arro3(py)?)
         }
         (AnyGeometryInput::Chunked(left), AnyGeometryBroadcastInput::Chunked(right)) => {
             let result = FrechetDistance::frechet_distance(&left.as_ref(), &right.as_ref())?;
-            Ok(PyChunkedArray::from_arrays(result.chunks())?.to_arro3(py)?)
+            Ok(PyChunkedArray::from_array_refs(result.chunk_refs())?.to_arro3(py)?)
         }
         (AnyGeometryInput::Array(left), AnyGeometryBroadcastInput::Scalar(right)) => {
             let scalar = right.to_geo_line_string()?;
             let result = FrechetDistanceLineString::frechet_distance(&left.as_ref(), &scalar)?;
-            Ok(PyArray::from_array(result).to_arro3(py)?)
+            Ok(PyArray::from_array_ref(Arc::new(result)).to_arro3(py)?)
         }
         (AnyGeometryInput::Chunked(left), AnyGeometryBroadcastInput::Scalar(right)) => {
             let scalar = right.to_geo_line_string()?;
             let result = FrechetDistanceLineString::frechet_distance(&left.as_ref(), &scalar)?;
-            Ok(PyChunkedArray::from_arrays(result.chunks())?.to_arro3(py)?)
+            Ok(PyChunkedArray::from_array_refs(result.chunk_refs())?.to_arro3(py)?)
         }
         _ => Err(PyValueError::new_err("Unsupported input types.").into()),
     }
