+++ python/core/src/algorithm/geo/geodesic_area.rs
@@ -1,3 +1,5 @@
+use std::sync::Arc;
+
 use crate::error::PyGeoArrowResult;
 use crate::ffi::from_python::AnyGeometryInput;
 use geoarrow::algorithm::geo::GeodesicArea;
@@ -9,11 +11,11 @@ pub fn geodesic_perimeter(py: Python, input: AnyGeometryInput) -> PyGeoArrowResu
     match input {
         AnyGeometryInput::Array(arr) => {
             let out = arr.as_ref().geodesic_perimeter()?;
-            Ok(PyArray::from_array(out).to_arro3(py)?)
+            Ok(PyArray::from_array_ref(Arc::new(out)).to_arro3(py)?)
         }
         AnyGeometryInput::Chunked(arr) => {
             let out = arr.as_ref().geodesic_perimeter()?;
-            Ok(PyChunkedArray::from_arrays(out.chunks())?.to_arro3(py)?)
+            Ok(PyChunkedArray::from_array_refs(out.chunk_refs())?.to_arro3(py)?)
         }
     }
 }
