+++ python/core/src/error.rs
@@ -6,7 +6,9 @@ pub enum PyGeoArrowError {
     PyErr(PyErr),
     PyArrowError(pyo3_arrow::error::PyArrowError),
     PythonizeError(pythonize::PythonizeError),
+    #[cfg(feature = "async")]
     ObjectStoreError(object_store::Error),
+    #[cfg(feature = "async")]
     ObjectStorePathError(object_store::path::Error),
     SerdeJsonError(serde_json::Error),
     UrlParseError(url::ParseError),
@@ -19,7 +21,9 @@ impl From<PyGeoArrowError> for PyErr {
             PyGeoArrowError::PyErr(err) => err,
             PyGeoArrowError::PyArrowError(err) => err.into(),
             PyGeoArrowError::PythonizeError(err) => PyException::new_err(err.to_string()),
+            #[cfg(feature = "async")]
             PyGeoArrowError::ObjectStoreError(err) => PyException::new_err(err.to_string()),
+            #[cfg(feature = "async")]
             PyGeoArrowError::ObjectStorePathError(err) => PyException::new_err(err.to_string()),
             PyGeoArrowError::SerdeJsonError(err) => PyException::new_err(err.to_string()),
             PyGeoArrowError::UrlParseError(err) => PyException::new_err(err.to_string()),
@@ -45,12 +49,14 @@ impl From<pythonize::PythonizeError> for PyGeoArrowError {
     }
 }
 
+#[cfg(feature = "async")]
 impl From<object_store::Error> for PyGeoArrowError {
     fn from(other: object_store::Error) -> Self {
         Self::ObjectStoreError(other)
     }
 }
 
+#[cfg(feature = "async")]
 impl From<object_store::path::Error> for PyGeoArrowError {
     fn from(other: object_store::path::Error) -> Self {
         Self::ObjectStorePathError(other)
