+++ python/core/src/io/parquet/mod.rs
@@ -1,3 +1,9 @@
+#[cfg(feature = "async")]
+mod r#async;
+mod sync;
+
 pub mod options;
-pub mod reader;
-pub mod writer;
+
+#[cfg(feature = "async")]
+pub use r#async::{read_parquet_async, ParquetDataset, ParquetFile};
+pub use sync::{read_parquet, write_parquet, ParquetWriter};
