+++ python/core/src/lib.rs
@@ -106,37 +106,37 @@ fn _rust(_py: Python, m: &Bound<PyModule>) -> PyResult<()> {
     )?)?;
     m.add_function(wrap_pyfunction!(crate::table::geometry_col, m)?)?;
 
-    // IO
+    // Async IO
 
-    m.add_class::<crate::io::object_store::PyObjectStore>()?;
-    m.add_class::<crate::io::parquet::reader::ParquetFile>()?;
-    m.add_class::<crate::io::parquet::reader::ParquetDataset>()?;
+    #[cfg(feature = "async")]
+    {
+        m.add_class::<crate::io::object_store::PyObjectStore>()?;
+        m.add_class::<crate::io::parquet::ParquetFile>()?;
+        m.add_class::<crate::io::parquet::ParquetDataset>()?;
+
+        m.add_function(wrap_pyfunction!(
+            crate::io::flatgeobuf::read_flatgeobuf_async,
+            m
+        )?)?;
+        m.add_function(wrap_pyfunction!(crate::io::parquet::read_parquet_async, m)?)?;
+
+        m.add_function(wrap_pyfunction!(crate::io::postgis::read_postgis, m)?)?;
+        m.add_function(wrap_pyfunction!(crate::io::postgis::read_postgis_async, m)?)?;
+    }
+
+    // IO
 
     m.add_function(wrap_pyfunction!(crate::io::csv::read_csv, m)?)?;
     m.add_function(wrap_pyfunction!(crate::io::flatgeobuf::read_flatgeobuf, m)?)?;
-    m.add_function(wrap_pyfunction!(
-        crate::io::flatgeobuf::read_flatgeobuf_async,
-        m
-    )?)?;
     m.add_function(wrap_pyfunction!(crate::io::geojson::read_geojson, m)?)?;
     m.add_function(wrap_pyfunction!(
         crate::io::geojson_lines::read_geojson_lines,
         m
     )?)?;
-    m.add_function(wrap_pyfunction!(
-        crate::io::parquet::reader::read_parquet,
-        m
-    )?)?;
-    m.add_function(wrap_pyfunction!(
-        crate::io::parquet::reader::read_parquet_async,
-        m
-    )?)?;
-    m.add_function(wrap_pyfunction!(
-        crate::io::parquet::writer::write_parquet,
-        m
-    )?)?;
-    m.add_function(wrap_pyfunction!(crate::io::postgis::read_postgis, m)?)?;
-    m.add_function(wrap_pyfunction!(crate::io::postgis::read_postgis_async, m)?)?;
+    m.add_function(wrap_pyfunction!(crate::io::parquet::read_parquet, m)?)?;
+    m.add_function(wrap_pyfunction!(crate::io::parquet::write_parquet, m)?)?;
+    m.add_class::<crate::io::parquet::ParquetWriter>()?;
+
     m.add_function(wrap_pyfunction!(
         crate::interop::pyogrio::from_pyogrio::read_pyogrio,
         m
